Implemented Write System Call
Implemented taint querying for the write system call. Successfully
tested tainting multiple sources, having their data be written to a
single sink, and querying the sink to correctly identify which sources'
tainted data was written to the sink.

+++ dependency_tracker/dependency_tracker_targets.cpp
@@ -43,7 +43,7 @@ const size_t& TargetSink::getIndex() const {
 	return this->index;
 }
 
-std::map<size_t, uint32_t>& TargetSink::getLabeledBytes() {
+std::map<uint32_t, uint32_t>& TargetSink::getLabeledBytes() {
 	return this->labeledBytes;
 }
